taigaProject/requirements.txt
requests>=2.26.1
python-dotenv~=1.0.0